requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
joblib
supabase
python-dotenv
openai
lime